requirements.txt
asgiref==3.7.2
Django==5.0.1
django-filter==23.5
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
